Fix error-prone issues part6

+++ constraintlayout/core/src/main/java/androidx/constraintlayout/core/state/ConstraintSetParser.java
@@ -220,7 +220,7 @@ public class ConstraintSetParser {
             int maxInt = (int) mMax;
             for (int i = value; i <= maxInt; i++) {
                 array.add(mPrefix + value + mPostfix);
-                value += (float) mStep;
+                value += (int) mStep;
             }
             return array;
